+++ resources/views/components/icon/logos/alpinejs.blade.php
@@ -1,0 +1,1 @@
+<svg {{ $attributes }} xmlns="http://www.w3.org/2000/svg" width="52.325" height="24" viewBox="0 0 52.325 24"><g transform="translate(-9.968 -10.303)"><path d="M121.276,10.3,132.9,21.88,121.276,33.456,109.648,21.88Z" transform="translate(-70.611)" fill="#77c1d2" fill-rule="evenodd"/><path d="M21.6,10.3,45.7,34.3H22.446L9.968,21.88Z" fill="#2d3441" fill-rule="evenodd"/></g></svg>