+++ resources/views/layouts/base.blade.php
@@ -10,6 +10,14 @@
 
     </head>
 
+    <script async src="https://www.googletagmanager.com/gtag/js?id={{ config('services.analytics_id') }}"></script>
+    <script>
+        window.dataLayer = window.dataLayer || [];
+        function gtag() { dataLayer.push(arguments); }
+        gtag('js', new Date());
+        gtag('config',  '{{ config('services.analytics_id') }}', {cookie_flags: 'SameSite=None;Secure'});
+    </script>
+
     <body {{ $attributes }}>
 
         {{ $slot }}
@@ -22,4 +30,4 @@
 
     </body>
 
-</html>
+    </html>