+++ stubs/config/services.php
@@ -10,7 +10,15 @@ return [
      *
      */
     'recaptcha_site_key' => env('RECAPTCHA_SITE_KEY', ''),
-    'recaptcha_secret_key' => env('RECAPTCHA_SECRET_KEY', '')
+    'recaptcha_secret_key' => env('RECAPTCHA_SECRET_KEY', ''),
+
+    /**
+     * ----------------------------------------------------------------------
+     * Google Analytics Stream Details
+     * ----------------------------------------------------------------------
+     *
+     */
+    'analytics_id' => env('ANALYTICS_ID', ''),
 
 
 ];